+++ src/http/http.js
@@ -5,12 +5,12 @@ var qs = require( 'qs' );
 var queryparse = qs.parse;
 var express = require( 'express' );
 var http = require( 'http' );
-var debug = require( 'debug' )( 'autohost:http-transport' );
+var log = require( '../log' )( 'autohost.http.transport' );
 var regex = require( './regex.js' );
 var Router = express.Router;
 var expreq = express.request;
 var expres = express.response;
-var middleware, userMiddleware, routes, paths, request, config, metrics, middlewareLib;
+var middleware, userMiddleware, routes, paths, request, config, middlewareLib;
 
 var wrapper;
 reset();
@@ -19,7 +19,7 @@ function buildUrl() {
 	var idx = 0,
 		cleaned = [],
 		segment;
-	while (idx < arguments.length) {
+	while ( idx < arguments.length ) {
 		segment = arguments[ idx ];
 		if ( segment.substr( 0, 1 ) === '/' ) {
 			segment = segment.substr( 1 );
@@ -173,7 +173,7 @@ function queryParser( req, res, next ) {
 
 function registerMiddleware( filter, callback ) {
 	var fn = function( router ) {
-		debug( 'MIDDLEWARE: %s mounted at %s', ( callback.name || 'anonymous' ), filter );
+		log.debug( 'MIDDLEWARE: %s mounted at %s', ( callback.name || 'anonymous' ), filter );
 		router.use( filter, callback );
 	};
 	if ( wrapper.app ) {
@@ -184,7 +184,7 @@ function registerMiddleware( filter, callback ) {
 
 function registerUserMiddleware( filter, callback ) {
 	var fn = function( router ) {
-		debug( 'MIDDLEWARE: %s mounted at %s', ( callback.name || 'anonymous' ), filter );
+		log.debug( 'MIDDLEWARE: %s mounted at %s', ( callback.name || 'anonymous' ), filter );
 		router.use( filter, callback );
 	};
 	if ( wrapper.app ) {
@@ -196,17 +196,15 @@ function registerUserMiddleware( filter, callback ) {
 function registerRoute( url, method, callback ) {
 	method = method.toLowerCase();
 	method = method === 'all' || method === 'any' ? 'all' : method;
-	var errors = [ 'autohost', 'errors', method.toUpperCase() + ' ' + url ].join( '.' );
 	var fn = function() {
 		url = prefix( url );
-		debug( 'ROUTE: %s %s -> %s', method, url, ( callback.name || 'anonymous' ) );
+		log.debug( 'ROUTE: %s %s -> %s', method, url, ( callback.name || 'anonymous' ) );
 		wrapper.app[ method ]( url, function( req, res ) {
 			if ( config && config.handleRouteErrors ) {
 				try {
 					callback( req, res );
-				} catch (err) {
-					metrics.meter( errors ).record();
-					debug( 'ERROR! route: %s %s failed with %s', method.toUpperCase(), url, err.stack );
+				} catch ( err ) {
+					log.debug( 'ERROR! route: %s %s failed with %s', method.toUpperCase(), url, err.stack );
 					res.status( 500 ).send( 'Server error at ' + method.toUpperCase() + ' ' + url );
 				}
 			} else {
@@ -227,7 +225,7 @@ function registerStaticPath( url, opt ) {
 	var fn = function() {
 		url = prefix( url );
 		var target = path.resolve( filePath );
-		debug( 'STATIC: %s -> %s', url, target );
+		log.debug( 'STATIC: %s -> %s', url, target );
 		wrapper.app.use( url, express.static( target, options ) );
 	};
 	paths.push( fn );
@@ -289,11 +287,10 @@ function reset() {
 	paths = [];
 }
 
-module.exports = function( req, mw, metric, resetState ) {
+module.exports = function( req, mw, resetState ) {
 	if ( resetState ) {
 		reset();
 	}
-	metrics = metric;
 	request = req;
 	middlewareLib = mw;