+++ src/http/passport.js
@@ -1,15 +1,12 @@
 var _ = require( 'lodash' );
 var when = require( 'when' );
 var passport = require( 'passport' );
-var debug = require( 'debug' )( 'autohost:passport' );
+var log = require( '../log' )( 'autohost.passport' );
+var metronic = require( '../metrics' );
 var noOp = function() {
 	return when( true );
 };
 var userCountCheck = noOp;
-var authorizationErrorCount = 'autohost.authorization.errors';
-var authorizationErrorRate = 'autohost.authorization.error.rate';
-var authenticationTimer = 'autohost.authentication.timer';
-var authorizationTimer = 'autohost.authorization.timer';
 var passportInitialize;
 var passportSession;
 var authProvider;
@@ -18,15 +15,17 @@ var metrics;
 
 reset();
 
-function authConditionally( req, res, next ) {  
+function authConditionally( req, res, next ) {
 	// if previous middleware has said to skip auth OR
 	// a user was attached from a session, skip authenticating
 	if ( req.skipAuth || req.user ) {
+		metrics.authenticationSkips.record();
 		next();
 	} else {
-		metrics.timer( authenticationTimer ).start();
+		metrics.authenticationAttempts.record();
+		var timer = metrics.authenticationTimer();
 		authProvider.authenticate( req, res, next );
-		metrics.timer( authenticationTimer ).record();
+		timer.record();
 	}
 }
 
@@ -44,57 +43,65 @@ function getAuthMiddleware( uri ) {
 	return list;
 }
 
-function getRoles( req, res, next ) {  
+function getRoles( req, res, next ) {
 	var userName = _.isObject( req.user.name ) ? req.user.name.name : req.user.name;
+	req.authenticatedUser = userName || req.user.id || 'anonymous';
+
+	function onError( err ) {
+		metrics.authorizationErrors.record();
+		req.user.roles = [];
+		timer.record();
+		log.debug( 'Failed to get roles for %s with %s', getUserString( req.user ), err.stack );
+		// during a socket connection, express is not fully initialized and this call fails ... hard
+		try {
+			res.status( 500 ).send( 'Could not determine user permissions' );
+		} catch ( err ) {
+			log.warn( 'Could not reply with user permission error to request before express is fully initialized.' );
+		}
+		next();
+	}
+
+	function onRoles( roles ) {
+		log.debug( 'Got roles [ %s ] for %s', roles, req.user );
+		req.user.roles = roles;
+		timer.record();
+		next();
+	}
+
 	if ( userName === 'anonymous' ) {
 		req.user.roles = [ 'anonymous' ];
 		next();
 	} else {
-		metrics.timer( authorizationTimer ).start();
+		var timer = metrics.authorizationTimer();
 		authProvider.getUserRoles( req.user, req.context )
-			.then( null, function( err ) {
-				metrics.counter( authorizationErrorCount ).incr();
-				metrics.meter( authorizationErrorRate ).record();
-				metrics.timer( authorizationTimer ).record();
-				debug( 'Failed to get roles for %s with %s', getUserString( req.user ), err.stack );
-				// during a socket connection, express is not fully initialized and this call fails ... hard
-				try {
-					res.status( 500 ).send( 'Could not determine user permissions' );
-				} catch (err) {
-					return [];
-				}
-			} )
-			.then( function( roles ) {
-				debug( 'Got roles [ %s ] for %s', roles, req.user );
-				req.user.roles = roles;
-				metrics.timer( authorizationTimer ).record();
-				next();
-			} );
+			.then( onRoles, onError );
 	}
 }
 
 function getSocketRoles( user ) {
+	function onError( err ) {
+		metrics.authorizationErrors.record();
+		timer.record();
+		log.debug( 'Failed to get roles for %s with %s', getUserString( user ), err.stack );
+		return [];
+	}
+
+	function onRoles( roles ) {
+		log.debug( 'Got roles [ %s ] for %s', roles, getUserString( user ) );
+		timer.record();
+		return roles;
+	}
+
 	if ( user.name === 'anonymous' ) {
 		return when( [ 'anonymous' ] );
 	} else {
-		metrics.timer( authorizationTimer ).start();
+		var timer = metrics.authorizationTimer();
 		return authProvider.getUserRoles( user, {} )
-			.then( null, function( err ) {
-				metrics.counter( authorizationErrorCount ).incr();
-				metrics.meter( authorizationErrorRate ).record();
-				metrics.timer( authorizationTimer ).record();
-				debug( 'Failed to get roles for %s with %s', getUserString( user ), err.stack );
-				return [];
-			} )
-			.then( function( roles ) {
-				debug( 'Got roles [ %s ] for %s', roles, getUserString( user ) );
-				metrics.timer( authorizationTimer ).record();
-				return roles;
-			} );
+			.then( onRoles, onError );
 	}
 }
 
-function getUserString( user ) {  
+function getUserString( user ) {
 	return user.name ? user.name : JSON.stringify( user );
 }
 
@@ -103,17 +110,16 @@ function reset() {
 	passportSession = passport.session();
 	authProvider = undefined;
 	anonPaths = undefined;
-	metrics = undefined;
 }
 
 function resetUserCount() {
 	userCountCheck = authProvider.hasUsers;
 }
 
-function skipAuthentication( req, res, next ) {  
+function skipAuthentication( req, res, next ) {
 	req.skipAuth = true;
 	if ( !req.user ) {
-		debug( 'Skipping authentication and assigning user anonymous to request %s %s', req.method, req.url );
+		log.debug( 'Skipping authentication and assigning user anonymous to request %s %s', req.method, req.url );
 		req.user = {
 			id: 'anonymous',
 			name: 'anonymous',
@@ -123,7 +129,7 @@ function skipAuthentication( req, res, next ) {
 	next();
 }
 
-function whenNoUsers( req, res, next ) {  
+function whenNoUsers( req, res, next ) {
 	userCountCheck()
 		.then( function( hasUsers ) {
 			if ( hasUsers ) {
@@ -142,11 +148,11 @@ function withAuthLib( authProvider ) {
 	} );
 }
 
-module.exports = function( config, authPlugin, meter, resetState ) {
+module.exports = function( config, authPlugin, resetState ) {
 	if ( resetState ) {
 		reset();
 	}
-	metrics = meter;
+	metrics = metronic();
 	authProvider = authPlugin;
 	authProvider.initPassport( passport );
 	passport.serializeUser( authProvider.serializeUser );