+++ src/websocket/socketio.js
@@ -1,6 +1,6 @@
 var _ = require( 'lodash' );
 var socketio = require( 'socket.io' );
-var debug = require( 'debug' )( 'autohost:socketio' );
+var log = require( '../log' )( 'autohost.socketio' );
 var authStrategy;
 var registry;
 var config;
@@ -8,7 +8,7 @@ var io;
 var middleware;
 
 function acceptSocket( socket ) {
-	debug( 'Processing socket.io connection attempt' );
+	log.debug( 'Processing socket.io connection attempt' );
 
 	var request = socket.request;
 	socket.type = 'socketio';
@@ -51,7 +51,7 @@ function acceptSocket( socket ) {
 
 	// add a way to close a socket
 	socket.close = function() {
-		debug( 'Closing socket.io client (user: %s)', JSON.stringify( socket.user ) );
+		log.debug( 'Closing socket.io client (user: %s)', JSON.stringify( socket.user ) );
 		socket.removeAllListeners();
 		socket.disconnect( true );
 		registry.remove( socket );
@@ -65,7 +65,7 @@ function acceptSocket( socket ) {
 
 	// if client identifies itself, register id
 	socket.on( 'client.identity', function( data ) {
-		debug( 'Client sent identity %s', JSON.stringify( data ) );
+		log.debug( 'Client sent identity %s', JSON.stringify( data ) );
 		socket.id = data.id;
 		registry.identified( data.id, socket );
 	} );
@@ -84,7 +84,7 @@ function acceptSocket( socket ) {
 
 	socket.publish( 'server.connected', { user: socket.user } );
 	socket.on( 'disconnect', function() {
-		debug( 'socket.io client disconnected (user: %s)', JSON.stringify( socket.user ) );
+		log.debug( 'socket.io client disconnected (user: %s)', JSON.stringify( socket.user ) );
 		socket.removeAllListeners();
 		registry.remove( socket );
 	} );
@@ -95,16 +95,16 @@ function authSocketIO( req, allow ) {
 	if ( authStrategy ) {
 		middleware
 			.use( '/', function( hreq, hres, next ) {
-				debug( 'Setting socket.io connection user to %s', JSON.stringify( hreq.user ) );
+				log.debug( 'Setting socket.io connection user to %s', JSON.stringify( hreq.user ) );
 				allowed = hreq.user;
 				next();
 			} )
 			.handle( req, req.res, function( err ) {
 				if ( err ) {
-					debug( 'Error in authenticating socket.io connection %s', err.stack );
+					log.debug( 'Error in authenticating socket.io connection %s', err.stack );
 					allow( err );
 				} else {
-					debug( 'Authenticated socket.io connection as user %s', allowed );
+					log.debug( 'Authenticated socket.io connection as user %s', allowed );
 					allow( null, allowed );
 				}
 			} );